requirements.txt
gradio==4.13.0
gradio-client==0.8.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.3
ipykernel==6.29.2
streamlit
matplotlib==3.7.4
matplotlib-inline==0.1.6
numpy
pandas
Pillow==9.5.0
tokenizers
torch
torchmetrics
torchtext
torchvision
transformers